`tools.utils` is now a package
