+++ tools/pkg/repo/create.py
@@ -19,12 +19,11 @@ from ptscripts import Context, command_group
 
 import tools.pkg
 import tools.utils
-from tools.utils import (
-    Version,
+from tools.utils import Version, parse_versions
+from tools.utils.repo import (
     create_full_repo_path,
     create_top_level_repo_path,
     get_repo_json_file_contents,
-    parse_versions,
 )
 
 try:
