+++ tools/pkg/repo/publish.py
@@ -20,13 +20,9 @@ from ptscripts import Context, command_group
 
 import tools.pkg
 import tools.utils
-from tools.utils import (
-    Version,
-    create_full_repo_path,
-    get_repo_json_file_contents,
-    get_salt_releases,
-    parse_versions,
-)
+import tools.utils.repo
+from tools.utils import Version, get_salt_releases, parse_versions
+from tools.utils.repo import create_full_repo_path, get_repo_json_file_contents
 
 try:
     import boto3
@@ -336,7 +332,7 @@ def release(ctx: Context, salt_version: str):
                         Bucket=bucket_name,
                         Key=path,
                         Fileobj=wfh,
-                        Callback=tools.utils.UpdateProgress(progress, task),
+                        Callback=tools.utils.repo.UpdateProgress(progress, task),
                     )
                 updated_contents = re.sub(
                     r"^(baseurl|gpgkey)=https://([^/]+)/(.*)$",
@@ -366,7 +362,7 @@ def release(ctx: Context, salt_version: str):
                         str(upload_path),
                         tools.utils.RELEASE_BUCKET_NAME,
                         str(relpath),
-                        Callback=tools.utils.UpdateProgress(progress, task),
+                        Callback=tools.utils.repo.UpdateProgress(progress, task),
                     )
 
 
@@ -438,7 +434,7 @@ def github(
                 Bucket=tools.utils.STAGING_BUCKET_NAME,
                 Key=str(entry_path),
                 Fileobj=wfh,
-                Callback=tools.utils.UpdateProgress(progress, task),
+                Callback=tools.utils.repo.UpdateProgress(progress, task),
             )
 
     for artifact in artifacts_path.iterdir():
@@ -618,7 +614,7 @@ def _publish_repo(
                     str(upload_path),
                     bucket_name,
                     str(relpath),
-                    Callback=tools.utils.UpdateProgress(progress, task),
+                    Callback=tools.utils.repo.UpdateProgress(progress, task),
                     ExtraArgs={
                         "Metadata": {
                             "x-amz-meta-salt-release-version": salt_version,
